requirements.txt
paho-mqtt
soco
